requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Limiter==3.5.0
bcrypt==4.0.1
python-dotenv==1.0.0
email-validator==2.0.0
cryptography==41.0.4
WTForms==3.0.1
gunicorn==21.2.0